requirements.txt
tensorflow==2.13.0
pandas 
gdown
dvc
mlflow==2.2.2
notebook
numpy
matplotlib
seaborn
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-PyYAML
scipy
Flask
Flask-Cors
openai
streamlit
python-dotenv
-e .